backend/requirements.txt
# Web Framework and Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database and Validation
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0

# Security and Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.0.0
joblib>=1.3.0

# Utilities
python-dotenv>=1.0.0